requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.4.1
Flask-SocketIO==3.3.2
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
MarkupSafe==1.1.1
passlib==1.7.1
python-engineio==3.5.1
python-socketio==3.1.2
python-barcode==0.13.1
